+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 ## PyMC 3.3. (Unreleased)
 
 - Improve NUTS initialization `advi+adapt_diag_grad` and add `jitter+adapt_diag_grad` (#2643)
+- Fixed `compareplot` to use `loo` output.
 
 
 ## PyMC3 3.2 (October 10, 2017)